+++ wazo_confd/config.py
@@ -1,4 +1,4 @@
-# Copyright 2013-2021 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2013-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import argparse
@@ -154,6 +154,7 @@ DEFAULT_CONFIG = {
         'user_line': True,
         'user_line_associated': True,
         'user_schedule': True,
+        'user_subscription': True,
         'user_voicemail': True,
         'voicemail': True,
         'voicemail_general': True,