+++ wazo_confd/plugins/user_subscription/plugin.py
@@ -1,0 +1,18 @@
+# Copyright 2022 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .resource import UserSubscription
+from .service import build_service
+
+
+class Plugin:
+    def load(self, dependencies):
+        api = dependencies['api']
+        service = build_service()
+
+        api.add_resource(
+            UserSubscription,
+            '/users/subscriptions',
+            endpoint='user_subscription',
+            resource_class_args=(service,),
+        )